Fix support for multiple Mesos Clouds
This is to address issue 94:
https://github.com/jenkinsci/mesos-plugin/issues/94

The Mesos object was written to be a singleton
even though Jenkins supports defining more than one.

In my case, I was trying to control 2 different clusters
(with 2 different slave tags).  Because the MesosImpl gets created by
whichever slave happens to launch first, slave tags from the other cluster
would launch in the wrong place.

While I don't LOVE this implementation, it seemed the least minimal
impact to a larger refactoring.  I basically pass a reference for the
MesosCloud to the MesosSlave.  Then I changed the Mesos.getInstance()
to take a parameter which is used as a key in a Map keyed by MesosCloud to
find the correct MesosImpl.  It was necessary to externalize the MesosImpl
inner-class to get it to compile.

I also cleaned up some unused imports, code not called, and depricated
use of Hudson for Jenkins.  You also can't override interfaces so I removed
some @Override tags