Add configurable network mode for docker containers
This commit adds the possibility to configure the network mode of the
containerized Jenkins slaves provisioned by Mesos.

For now it is possible to pick "Bridge" or "Host", since these are
the two supported modes by Mesos (version 0.21.x).

Additionally, in "Bridge" mode it is also possible to define port mappings.

Issue: JENKINS-26324