Add jenkins label to mesos slave task's name
When running many slave types, seeing everything as:

    mesos-jenkins-453c7f5c-4c70-46fa-a272-cf939852f321
    mesos-jenkins-7e2a8b6a-66e0-4cbc-87ec-cf06a4693a7e
    mesos-jenkins-06e6dd76-20e0-4b69-8fef-47d039d822f0

It becomes difficult to see what is what.

This change adds the jenkins label into the name so it looks more like this:

    mesos-jenkins-453c7f5c4c7046faa272cf939852f321-foo-slave
    mesos-jenkins-7e2a8b6a66e04cbc87eccf06a4693a7e-bar-slave
    mesos-jenkins-06e6dd7620e04b698fef47d039d822f0-baz-slave

where the jenkins labes are `foo-slave`, `bar-slave` and `baz-slave`

However, the docker image won't start if the name is too long.
RFC1034/1035 recommends limiting this to 63 characters max.
Rather than trimming the UUID, we trim the slave name at the end to
avoid collisions.  But we remove the internal dashes in the UUID
to save a few characters.