Fix another case of duplicate framework registration
Added a Garbage Collector to stop the scheduler and remove the mesos framework in case the cloud gets removed.

Doing so, I found out that there was a problem related to credentialsId and usage of @DataBoundSetter : when the scheduler is restarted after constructor call (onDemandRegistration=false), credentialsId value is not yet set and this can cause a duplicate framework registration. So the field initialization has been moved to constructor.

Last, added equals/hashCode implementations to MesosSlaveInfo and embedded classes, as these are necessary to compare MesosCloud instances, and these are used in Mesos#getInstance to prevent building duplicate MesosImpl instances.