Handle backpressure from commands queue.  (#46)
Summary:
When the USI commands queue is filled we should not provision more
agents. This change lets the agent provisioning failed since Jenkins
won't give us an option to backpressure.

This change also introduces also a `MotherObject` for `AgentSpecTemplates`.
This ease sharing fixtures. See the linked article in the docs.

Furthermore we introduce operational settings for the plugin that are the
same across all `MesosCloud` instances.

JIRA issues: DCOS_OSS-5054